+++ tests/eval/regression/known_pending_rules.py
@@ -35,22 +35,16 @@ KNOWN_PENDING_RULES: frozenset[tuple[str, str]] = frozenset(
         ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-malware-2"),  # nx_enabled
         ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-malware-3"),  # image_profile_acceptance
         ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-server-1"),  # lockdown_mode
-        ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-server-2"),  # ssh_enabled
         ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-server-3"),  # ad_joined
-        ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-server-4"),  # ntp_servers
         ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-server-6"),  # patch_status
         ("bsi-itgs-basisabsicherung-vmware", "bsi-itgs-server-7"),  # secure_boot_enabled
         # --- cis-vmware-esxi-8.0-subset ---
-        ("cis-vmware-esxi-8.0-subset", "cis-esxi-2.1.1"),  # ntp_enabled
-        ("cis-vmware-esxi-8.0-subset", "cis-esxi-2.1.2"),  # ntp_service_policy_on
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-2.2.2"),  # lockdown_mode_enabled
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-3.1.2"),  # persistent_logs
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-3.1.3"),  # audit_retention_days
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-4.1.1"),  # mgmt_vmk_isolated
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-4.1.2"),  # vswitch_promiscuous_mode
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-4.1.3"),  # forged_transmits
-        ("cis-vmware-esxi-8.0-subset", "cis-esxi-5.1.1"),  # firewall_enabled
-        ("cis-vmware-esxi-8.0-subset", "cis-esxi-5.1.2"),  # ssh_running
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-6.1.1"),  # ad_joined
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-6.1.2"),  # lockdown_exceptions_count
         ("cis-vmware-esxi-8.0-subset", "cis-esxi-6.1.3"),  # root_ssh_key_auth
@@ -64,11 +58,9 @@ KNOWN_PENDING_RULES: frozenset[tuple[str, str]] = frozenset(
         ("dengbao-2.0-level3-vmware", "db-l3-host-1"),  # root_ssh_key_auth
         ("dengbao-2.0-level3-vmware", "db-l3-host-2"),  # ad_joined
         ("dengbao-2.0-level3-vmware", "db-l3-host-3"),  # lockdown_mode_enabled
-        ("dengbao-2.0-level3-vmware", "db-l3-host-4"),  # firewall_enabled
         ("dengbao-2.0-level3-vmware", "db-l3-host-6"),  # audit_retention_days
         ("dengbao-2.0-level3-vmware", "db-l3-host-7"),  # host_secure_boot
         ("dengbao-2.0-level3-vmware", "db-l3-host-8"),  # host_tpm_attested
-        ("dengbao-2.0-level3-vmware", "db-l3-host-9"),  # ntp_enabled, ntp_service_policy_on
         ("dengbao-2.0-level3-vmware", "db-l3-vm-2"),  # secure_boot
         ("dengbao-2.0-level3-vmware", "db-l3-vm-3"),  # encryption_enabled, tags
         ("dengbao-2.0-level3-vmware", "db-l3-data-1"),  # encryption_enabled
@@ -86,7 +78,6 @@ KNOWN_PENDING_RULES: frozenset[tuple[str, str]] = frozenset(
         ("eu-nis2-vmware", "nis2-sc-1"),  # secure_boot_enabled
         # --- pci-dss-4.0-vmware ---
         ("pci-dss-4.0-vmware", "pci-r2-1"),  # default_admin_disabled
-        ("pci-dss-4.0-vmware", "pci-r2-2"),  # ssh_enabled
         ("pci-dss-4.0-vmware", "pci-r7-1"),  # ad_joined
         ("pci-dss-4.0-vmware", "pci-r8-1"),  # root_ssh_key_auth_enabled
         ("pci-dss-4.0-vmware", "pci-r8-2"),  # lockdown_mode